requirements.txt
# Core ML libraries
torch>=2.2.0
pytorch-lightning>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data processing
pyarrow>=14.0.0
pyyaml>=6.0

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Baseline models
xgboost>=2.0.0
catboost>=1.2.0
lifetimes>=0.11.3

# Logging and monitoring
wandb>=0.16.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Visualization (for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# SHAP for model interpretation
shap>=0.42.0 